+++ resources/views/patients/preview_patient.blade.php
@@ -15,7 +15,7 @@
                 </a>
             </div>
             <div class="col-md-auto">
-                <a href="/patients/delete_patient/{{$patient->id}}" class="text-danger">
+                <a href="/patients/delete_patient/{{$patient->id}}" class="text-danger" onclick="return confirm('Are you sure?')">
                     <i class="fas fa-times fs-3"></i><br>
                     <strong>Del</strong>
                 </a>
